gui: Sort language names vertically and don't truncate on small screens
Currently, language names are always displayed in two columns and sorted
from left to right first, and from top to bottom next. This commit
changes the sort order to go from top to bottom first, and only then
move to the next column.

Also, on small screens, display the languages in a single column only,
which allows for language names not to be truncated as much as they are
right now.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>