cmd/syncthing: Implement a subcommand 'override-credentials' (fixes #8118)
The subcommand takes one argument --src-dir which specifies the source
directory containing key.pem, cert.pem and config.xml files.
Then it replaces the key, cert and config files in the default folder
(.config/syncthing) with the ones from the source directory, which
gives the device a new identity.

The subcommand validates the input files which makes it safer than
replacing the files manually.