gui: Let folder ID be editable when adding folder (fixes #8145)
Somewhere in the recent refactors this got broken so that the folder ID
is not editable. I think it's because it shouldn't be editable when
adding a folder from the folder-rejected prompt. The template checks for
`edit` and `add` and makes the control read only, so I added a new value
`new` for this case.