gui: Add default ignores to the advanced configuration editor modal.
Follow the XML / JSON structure and keep it under a separate headline
instead of the "default folder" section (as in the regular settings
modal).

This section slightly differs in the handling of list-type config
options.  Usually they are represented as comma-separated single-line
text inputs, which is unsuitable for ignore patterns which may
rightfully contain commas.  Thus a textarea is used, using newline as
the list delimiter.