gui: Add detailed file versioning information to folder info (ref #963)
Currently, only the type of file versioning is displayed to the user in
the folder info. This is often not enough, especially when the user has
many folders with customised versioning settings that they would like to
be able to inspect quickly using the GUI.

This commit adds information about all user-controlled file versioning
settings right to the folder info, so that they can be examined without
editing each folder separately. In order not to make the GUI needlessly
more complex than it already is, the additional information is displayed
only when the fields use non-default values.

Specifically, the additional information contains:

1. "Clean out after" for Trash Can and Simple versioning
2. "Keep Versions" for Simple versioning
3. "Maximum Age" for Staggered versioning
4. "Cleanup Interval" and "Versions Path" for all but External
    versioning
5. "Command" for External versioning

The displayed information is condensed as much as possible, so that it
takes very little space vertically. Each value also has a tooltip that
displays additional explanatory information about it.

In addition, we remove the superfluous "File Versioning" text after each
versioning type, so that "Trash Can File Versioning" is shortened to
just "Trash Can" and so on. The text is redundant, because the exact
same "File Versioning" is already present in the left column of the
folder info table.

This commit also fixes the duration filter JS script, so that it does
not add reduntant spaces in front of the duration string, but only do
it in between the different time values.

Lastly, as a cosmetic update, the commit also updates the folder info
file versioning icon, so that it uses the black version of the icon,
which matches the other icons in the left column of the folder info
table better. While doing so, we also update the outdated code syntax.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>