gui: Display "days" next to "maximum age" in Staggered Versioning
Currently, the user is supposed to set the maximum age for Staggered
Versioning, with its default value being "365". However, there is no
indication what the value actually corresponds to.

This commit adds the text "days" displayed right to the input value,
which makes it clearer what the number actually means. This makes the
code identical to the clear out after input field in Simple Versioning
that also displays "days" next to its value.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>