all: Support syncing ownership (fixes #xxx)
This adds support for syncing ownership on Unixes and on Windows. The
scanner always picks up ownership information, but it is not applied
unless the new folder option "Sync Ownership" is set.

Ownership data is stored in a new FileInfo field called "OS data". This
is intended to hold further OS specific data in the future
(specifically, extended attributes), which is why the whole design is a
bit overkill for just ownership.