Set excludeCredentials in frontend instead of backend for consistency with UI state
Before, it was easy for backend state and UI state to end up with
conflicts. For example, if you delete a credential in the frontend and
attempt to register the same authenticator again before saving, the
backend would still include the deleted credential in
`excludeCredentials` and the authenticator wouldn't be able to be
registered again.

By instead setting `excludeCredentials` in the frontend, the above
scenario is no longer a problem because the `excludeCredentials`
setting is consistent with what's shown in the UI.

There is still a remaining state conflict, though: if you cancel the
settings dialog without saving, any credentials registered will remain
registered even though they "should" be removed because the UI state
wasn't explicitly saved.