gui: Make QR code image downloadable on click
Currently, the QR code is displayed as a simple image. This means that
if the user wants to download it, they need to right click (or long
press) it to display the menu first, and only the save it as image.

This commit transforms the image into a clickable link that presents the
user with a download dialog window on click. This way, it is easier to
download the image by non-tech-savvy users, and also the image filename
is automatically set to the actual device ID instead of being called
"download".

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>