gui: Allow to switch between folders/devices from inside Edit modal (fixes #7737)
Currently, one needs to repeatedly open and close the edit modal per
folder/device in order to check or modify settings of multiple folders
or devices. This process is slow and cumbersome. Thus, add "previous"
and "next" buttons directly inside the edit folder/device, so that the
user can switch between them without having to exit the modal.

When doing the switching, we also make sure to preserve the currently
opened tab, so that the user can easily compare the same settings for
different folders/devices.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>