lib/model: Correctly handle xattrs on directories
I'd forgotten to add code to set xattrs to the directory handling paths.
To fix that and avoid similar mistakes in the future I moved the xattr
handling to a new adjustCommonMetadata function, which previously just
handled ownership. The idea is for this function to apply common things
to all file types, so we can add things in one place and not forget half
of them...